backend/requirements.txt
alembic==1.18.4
fastapi==0.111.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
Mako==1.3.10
orjson==3.11.7
priority==2.0.0
psycopg2-binary==2.9.7
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.1
SQLAlchemy==2.0.21
starlette==0.37.2
typer==0.20.0
typer-slim==0.20.0
ujson==5.11.0
uvicorn==0.23.1
wsproto==1.3.2
PyJWT==2.11.0
